refactor(web): steps 4-5 — extract command-event registry and background-task tracker to store/chat/
